requirements.txt
affine==2.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.1
annotated-types==0.7.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.3.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
coloredlogs==15.0.1
colorlog==6.9.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contextily==1.6.2
contourpy==1.3.0
cycler==0.12.1
debugpy @ file:///C:/b/abs_bf9oo2vhxp/croot/debugpy_1736269476451/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
eval_type_backport==0.2.2
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.3.2
gdown==5.2.0
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
greenlet==3.2.2
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgviz==1.7.6
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources==6.5.2
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1680185618122/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257377578/work
kiwisolver==1.4.7
labelme==5.8.1
lazy_loader==0.4
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mercantile==1.2.1
mpmath==1.3.0
natsort==8.4.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.2.1
numpy==2.0.2
onnxruntime==1.19.2
opencv-python==4.10.0
opencv-python-headless==4.10.0
optuna==4.3.0
osam==0.2.3
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.2.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
protobuf==6.31.0
psutil @ file:///C:/b/abs_b5gv3mn55h/croot/psutil_1736371546320/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pydantic==2.11.4
pydantic_core==2.33.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.6.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1666828546753/work
rasterio==1.4.3
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
simsimd==6.4.7
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
soupsieve==2.7
SQLAlchemy==2.0.41
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
stringzilla==3.12.5
sympy==1.14.0
tifffile==2024.8.30
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
torchvision==0.22.0
tornado @ file:///D:/bld/tornado_1666788767305/work
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata==2025.2
urllib3==2.4.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
win32_setctime==1.2.0
xyzservices==2025.4.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
